Move reason test from [initState] to [askingReplyRequestReason]
