+++ src/webhook/handlers/__fixtures__/askingReplyRequestReason.js
@@ -1,0 +1,21 @@
+export const apiError = {
+  errors: [{ message: 'Syntax error' }],
+};
+
+export const updateReplyRequestSuccess = {
+  data: {
+    CreateOrUpdateReplyRequest: {
+      text: 'article text',
+      replyRequestCount: 1,
+    },
+  },
+};
+
+export const updateReplyRequestSuccess2 = {
+  data: {
+    CreateOrUpdateReplyRequest: {
+      text: 'article text',
+      replyRequestCount: 2,
+    },
+  },
+};
