+++ src/webhook/handlers/__fixtures__/initState.js
@@ -181,25 +181,3 @@ export const notFound = {
     },
   },
 };
-
-export const apiError = {
-  errors: [{ message: 'Syntax error' }],
-};
-
-export const updateReplyRequestSuccess = {
-  data: {
-    CreateOrUpdateReplyRequest: {
-      text: 'article text',
-      replyRequestCount: 1,
-    },
-  },
-};
-
-export const updateReplyRequestSuccess2 = {
-  data: {
-    CreateOrUpdateReplyRequest: {
-      text: 'article text',
-      replyRequestCount: 2,
-    },
-  },
-};
